fix(qemu): Do not intercept timer interrupts in device tree
qemu now adds a fifth "hypervisor virtual timer IRQ"
to the device tree, which exposed a bug in our code
that intercepted the device tree. That code is no longer
needed anyway because Linux does not use the secure timer
IRQ anymore on ARM64. Remove the code altogether and
pass in the timer interrupts to Linux unmodified.

Bug: 430121095
Test: presubmit
Change-Id: I3bb9338b46ea2d71a6ef26d2f86d831efeeee240